+++ LzpCache.php
@@ -42,7 +42,7 @@
 *	//Parametros( = Padrão):
 *		$config['dir'] = __DIR__.'/cache/'; 										//Caminho do Diretório onde o cache será armazenado
 *		$config['expire'] = 600; 													//0 para infinito - Valor Aceito int(opcional)
-*		$config['version'] = false; 												//false ou 0 desativam - Valores Aceitos float, string e int(opcional)
+*		$config['version'] = null; 													//null desativa - Valores Aceitos float, string e int(opcional)
 *		$config['compress'] = 0;													//0 desativa - Valor Aceito int de 0 a 9(opcional)
 *		$config['cacheNameType'] = array('hash' => 'md5', 'prefix' => '%name%_'); 	//Use %name% para colocar o nome do cache no prefixo(opcional)
 *		$config['ext'] = '.lzp'; 													//Extensão do arquivo de cache(opcional)
@@ -56,14 +56,14 @@
 *	//Parametros( = Padrão):
 * 		$cacheName = 'nome_do_cache'; 									//Nome do cache(Parametro obrigatório)
 * 		$getExpired = false;											//Ignora se o cache já expirou(opcional)
-* 		$cacheVersion = false;											//Versão do cache a ser obtido - Valores Aceitos float, string e int(opcional)
+* 		$cacheVersion = null;											//Versão do cache a ser obtido - Valores Aceitos float, string e int(opcional)
 *
 * Para obter múltiplos caches:
 * 	$cache->Get($cachesNames, $getExpired, $cacheVersion);				//Retorna um array($nomeDoCache=>$valor)
 * 	//Parametros( = Padrão):
 *		$cachesNames = array('nome_do_cache00', 'nome_do_cache01');		//Array contendo o Nome de cada cache(Parametro obrigatório)
 *		$getExpired = false;											//Ignora se o cache já expirou(opcional)
-* 		$cacheVersion = false;											//Versão do cache a ser obtido - Valores Aceitos float, string e int(opcional)
+* 		$cacheVersion = null;											//Versão do cache a ser obtido - Valores Aceitos float, string e int(opcional)
 *
 *
 *
@@ -72,14 +72,14 @@
 *	//Parametros( = Padrão):
 * 		$cacheName = 'nome_do_cache';														//Nome do cache(Parametro obrigatório)
 * 		$data = 'dadosDoCache';																//Dados a serem guardados no cache, tudo é aceito(Parametro obrigatório)
-* 		$cacheVersion = false;																//Versão do cache a ser criado - Valores Aceitos(float, string, int) - (opcional)
+* 		$cacheVersion = null;																//Versão do cache a ser criado - Valores Aceitos(float, string, int) - (opcional)
 *		$config = false;																	//Configurações para o cache Aceito: array('expire', 'compress')
 *
 * Para criar múltiplos caches:
 *  	$cache->CreateMultiples($namesValues, $cacheVersion); 									//Retorna um array($nomecache=>$foiCriado)
 *	//Parametros( = Padrão):
 *		$namesValues = array('nome_do_cache00' => $value, 'nome_do_cache01' => $value); 	//Array contendo os Nomes e os valores dos caches a serem criados(Parametro obrigatório)
-*		$cacheVersion = false; 																//Versão dos caches a serem criados - Valores Aceitos float, string e int(opcional)
+*		$cacheVersion = null; 																//Versão dos caches a serem criados - Valores Aceitos float, string e int(opcional)
 *
 *
 *
@@ -87,31 +87,31 @@
 * 	$cache->Delete($cacheName, $cacheVersion); 							//Retorna true se o cache for excluido, false em caso de falha e null caso o cache não exista
 * 	//Parametros( = Padrão):
 * 		$cacheName = 'nome_do_cache'; 									//Nome do cache(Required)
-* 		$cacheVersion = false; 											//Versão do cache a ser deletado - Valores Aceitos float, string e int(Optional)
+* 		$cacheVersion = null; 											//Versão do cache a ser deletado - Valores Aceitos float, string e int(Optional)
 *
 * Para Deletar Múltiplos caches
 * 	$cache->Delete($cachesNames, $cacheVersion); 						//Retorna um array($nomecache=>$foiDeletado), $foiDeletado = true(sucesso), false(falha) ou null(cache não existe)
 * 	//Parametros( = Padrão):
 * 		$cachesNames = array('nome_do_cache00', 'nome_do_cache01'); 	//Array contendo o Nome de cada cache(Required)
-*		$cacheVersion = false; 											//Versão dos caches a serem deletados - Valores Aceitos float, string e int(Optional)
+*		$cacheVersion = null; 											//Versão dos caches a serem deletados - Valores Aceitos float, string e int(Optional)
 *
 * Para deletar todos os caches
 * 	$cache->DeleteAll($cacheVersion); 									//Retorna true se os caches forem excluídos
 * 	//Parametros( = Padrão):
-* 		$cacheVersion = false; 											//Deleta os caches de uma certa versão - Valores Aceitos float, string e int(Parametro opicional)
+* 		$cacheVersion = null; 											//Deleta os caches de uma certa versão - Valores Aceitos float, string e int(Parametro opicional)
 *
 *
 * Para verificar se um cache existe
 * 	$cache->Exists($cacheName, $cacheVersion);							//Retorna true se o cache existe
 * 	//Parametros( = Padrão):
 * 		$cacheName = 'nome_do_cache'; 									//Nome do cache(Required)
-*		$cacheVersion = false; 											//Versão a ser verificada - Valores Aceitos float, string e int(Optional)
+*		$cacheVersion = null; 											//Versão a ser verificada - Valores Aceitos float, string e int(Optional)
 *
 * Para verificar se vários caches existem
 * 	$cache->Exists($cachesNames, $cacheVersion);						//Retorna um array($nomecache=>$exists)
 * 	//Parametros( = Padrão):
 * 		$cachesNames = array('nome_do_cache00', 'nome_do_cache01'); 	//Array contendo o Nome de cada cache(Required)
-*		$cacheVersion = false; 											//Versão a ser verificada - Valores Aceitos float, string e int(Optional)
+*		$cacheVersion = null; 											//Versão a ser verificada - Valores Aceitos float, string e int(Optional)
 *
 *
 *
@@ -119,13 +119,14 @@
 *	$cache->Size($dir, $version);
 *	//Parametros( = Padrão):
 *		$dir = null;								//Diretório a ser verificado(Optional)	
-*		$cacheVersion = false; 						//Retorna o tamanho do cache de uma certa versão - Valores Aceitos float, string e int(Optional)
+*		$cacheVersion = null; 						//Retorna o tamanho do cache de uma certa versão - Valores Aceitos float, string e int(Optional)
 *
 *
 *
 *** ChangeLog ***
 #####################################################################
 # V 2.0.0															#
+# -Código Documentado												#
 # -Código Reescrito													#
 # -Performance Otimizada											#
 # -Removida Criptografia do cache									#
@@ -180,19 +181,26 @@
 #####################################################################
 */
 class Cache{
-	const VERSION = '2.0.0';
 	const DS = DIRECTORY_SEPARATOR;
 	
-	#######
-	# CFG #
-	#######
+    /**
+     * Configuração atual.
+     * 
+     * Array contendo as configurações atuais.
+     * 
+     * @var array
+     */
 	protected $cfg;
 
-	#############
-	# CONSTRUCT #
-	#############
-	function __construct($config = false){
-		$configDefault = array(
+    /**
+     * Inicia junto com a classe
+     * 
+     * @param array $config Array Opcional contendo configurações para o cache.
+     * @return void
+     * @throws $config é inválido.
+     */
+	function __construct($config = null){
+		$this->cfg = array(
 			'dir' => (__DIR__).self::DS.'cache'.self::DS, 
 			'expire' => 600, 
 			'compress' => 0, 
@@ -202,65 +210,38 @@ class Cache{
 			'ext' => '.lzp'
 		);
 		if(is_array($config))
-			$this->cfg = array_replace($configDefault, $config);
-		else
-			$this->cfg = $configDefault;
+			$this->cfg = array_replace($this->cfg, $config);
+		elseif(!is_null($config))
+			throw new Exception('$config não é um array.');
 
-		if(!is_dir($this->cfg['dir'])){
+		if(!is_dir($this->cfg['dir']))
 			mkdir($this->cfg['dir'], 0755, true);
-		}
 	}
 
-	#################################################
-	# CONFIG/NAME/ENCODE/DECODE/COMPRESS/UNCOMPRESS #
-	#################################################
+    /**
+     * Configura a classe
+     * 
+     * @param array $config Array contendo configurações para o cache.
+     * @return void
+     */
 	public function Config($config){
 		$this->cfg = array_replace($this->cfg, $config);
-		if(!is_dir($this->cfg['dir'])){
+		if(!is_dir($this->cfg['dir']))
 			mkdir($this->cfg['dir'], 0755, true);
-		}
 	}
 
-	private function Filter($name){
-		return preg_replace("/[^a-zA-Z0-9_.-]/", "", $name);
-	}
-
-	private function Name($name){
-		$name = $this->Filter(strtolower($name));
-		$name = str_ireplace('%name%', $name, $this->cfg['cacheNameType']['prefix']).hash($this->cfg['cacheNameType']['hash'], $name);
-
-		return $name;
-	}
-
-	private function Encode($data){
-		return (is_array($data) || is_object($data)) ? serialize($data) : $data;
-	}
-
-	private function Decode($data){
-		$x = @unserialize($data);
-		return ($x === 'b:0;' || $x !== false) ? $x : $data;
-	}
-
-	private function Compress($data, $compressLevel){
-		return (function_exists('gzdeflate') && function_exists('gzinflate')) ? gzdeflate($data, $compressLevel) : $data;
-	}
-
-	private function Uncompress($data){
-		return function_exists('gzinflate') ? gzinflate($data) : $data;
-	}
-
-	private function GetVersion($version){
-		$version = !is_null($version) ? $version : $this->cfg['version'];
-		return !is_null($version) ? $this->Filter($version) : '';
-	}
-
-	##########
-	# EXISTS #
-	##########
+    /**
+     * Verifica se um ou mais caches existem
+     * 
+     * @param mixed $names Nome do cache a ser obtido, ou um array contendo os nomes
+	 * @param mixed $version Versão do(s) cache(s) a ser(em) obtido(s)
+     * @return mixed
+     */
 	public function Exists($names, $version=null){
 		$version = $this->GetVersion($version);
 
 		if(is_array($names)){
+			$exists = array();
 			foreach($names as $name){
 				$file = $this->cfg['dir'].$version.$this->Name($name).$this->cfg['ext'];
 				$exists[$name] = (is_file($file) && is_readable($file));
@@ -277,9 +258,9 @@ class Cache{
 	# CREATE #
 	##########
 	public function Create($name, $data, $version=null, $config=null){
-		if(!is_writeable($this->cfg['dir'])){
+		if(!is_writeable($this->cfg['dir']))
 			die('Direrório não diponível ou sem permissão para escrita');
-		}
+
 		$name = $this->Name($name);
 		$data = $this->Encode($data);
 		$version = $this->GetVersion($version);
@@ -298,9 +279,8 @@ class Cache{
 
 		$cacheData = $this->Encode($cacheData);
 
-		if(!is_dir($path)){
+		if(!is_dir($path))
 			mkdir($path, 0775, true);
-		}
 
 		return file_put_contents($path.$name.$this->cfg['ext'], $cacheData);
 	}
@@ -309,9 +289,9 @@ class Cache{
 	# CREATE MULTIPLES #
 	####################
 	public function CreateMultiples($values, $version=null){
-		if(!is_writeable($this->cfg['dir'])){
+		if(!is_writeable($this->cfg['dir']))
 			die('Direrório não diponível ou sem permissão para escrita');
-		}
+
 		$compress = ($this->cfg['compress'] > 0 && $this->cfg['compress'] < 10) ? $this->cfg['compress'] : 0;
 		$expire = ($this->cfg['expire']!=0) ? (time() + (Int)$this->cfg['expire']) : 0;
 		$version = $this->GetVersion($version);
@@ -339,9 +319,8 @@ class Cache{
 		$data = null;
 		$version = $this->GetVersion($version);
 
-		if(!is_readable($this->cfg['dir'])){
+		if(!is_readable($this->cfg['dir']))
 			die('Direrório não diponível ou sem permissão para leitura');
-		}
 
 		if(is_array($names)){
 			foreach($names as $name){
@@ -376,20 +355,21 @@ class Cache{
 	# DELETE #
 	##########
 	public function Delete($names, $version=null){
-		if(!is_writeable($this->cfg['dir']) && !is_readable($this->cfg['dir']))return;
-		$del = null;
+		if(!is_writeable($this->cfg['dir']) && !is_readable($this->cfg['dir']))
+			return;
+
 		$version = $this->GetVersion($version);
 
 		if(is_array($names)){
+			$del = array();
 			foreach($names as $name){
 				$file = $this->cfg['dir'].$version.$this->Name($name).$this->cfg['ext'];
 				$del[$name] = is_file($file) ? @unlink($file) : null;
 			}
 		}else{
 			$file = $this->cfg['dir'].$version.$this->Name($name).$this->cfg['ext'];
-			if(is_file($file)){
+			if(is_file($file))
 				$del = @unlink($file);
-			}
 		}
 		return $del;
 	}
@@ -398,14 +378,14 @@ class Cache{
 	# DELETE ALL #
 	##############
 	public function DeleteAll($version=null){
-		if(!is_writeable($this->cfg['dir']) && !is_readable($this->cfg['dir']))return;
-		$del = null;
+		if(!is_writeable($this->cfg['dir']) && !is_readable($this->cfg['dir']))
+			return;
+		$del = array();
 
 		$files = glob($this->cfg['dir'].$this->GetVersion($version).'/*', GLOB_NOSORT);
 		foreach($files as $file){
-			if(is_file($file)){
+			if(is_file($file))
 				$del[] = @unlink($file);
-			}
 		}
 		return (!is_null($del) && !in_array(false, $del));
 	}
@@ -415,9 +395,9 @@ class Cache{
 	########
 	public function Size($dir=null, $version=null){
 		$dir = (!is_null($dir) ? $dir : $this->cfg['dir']).$this->GetVersion($version);
-		if(!is_readable($dir)){
+		if(!is_readable($dir))
 			die('Direrório não diponível ou sem permissão para leitura');
-		}
+
 		$size = 0;
 		$files = glob(rtrim($dir, '/').'/*', GLOB_NOSORT);
 		foreach($files as $file){
@@ -427,5 +407,81 @@ class Cache{
 		$extSize = [' B', ' KB', ' MB', ' GB', ' TB', ' PB', ' EB', ' ZB', ' YB'];
 
 		return ($size ? round($size/pow(1024, ($i = floor(log($size, 1024)))), 2) . $extSize[$i] : 'Vazio');
+	}
+
+    /**
+     * Filtra uma string
+     * 
+     * @param string $data String para ser filtrada, removendo qualquer caractere inválido.
+     * @return string
+     */
+	private function Filter($data){
+		return preg_replace("/[^a-zA-Z0-9_.-]/", "", $data);
+	}
+
+    /**
+     * Retorna o nome final do cache
+     * 
+     * @param string $name Nome do cache.
+     * @return string
+     */
+	private function Name($name){
+		$name = $this->Filter(strtolower($name));
+		$name = str_ireplace('%name%', $name, $this->cfg['cacheNameType']['prefix']).hash($this->cfg['cacheNameType']['hash'], $name);
+
+		return $name;
+	}
+
+    /**
+     * Codifica o cache
+     * 
+     * @param mixed $data Array ou Objeto que será convertido para string.
+     * @return string
+     */
+	private function Encode($data){
+		return (is_array($data) || is_object($data)) ? serialize($data) : $data;
+	}
+
+    /**
+     * Decodifica o cache
+     * 
+     * @param mixed $data string que será decodificada.
+     * @return mixed
+     */
+	private function Decode($data){
+		$x = @unserialize($data);
+		return ($x === 'b:0;' || $x !== false) ? $x : $data;
+	}
+
+    /**
+     * Comprime o cache
+     * 
+     * @param string $data string que será comprimida.
+	 * @param int $compressLevel nível de compressão de 1 a 9.
+     * @return string
+     */
+	private function Compress($data, $compressLevel){
+		return (function_exists('gzdeflate') && function_exists('gzinflate')) ? gzdeflate($data, $compressLevel) : $data;
+	}
+
+    /**
+     * Descomprime o cache
+     * 
+     * @param string $data string que será descomprimida.
+     * @return string
+     */
+	private function Uncompress($data){
+		return function_exists('gzinflate') ? gzinflate($data) : $data;
+	}
+
+    /**
+     * Retorna a versão do cache filtrada
+     * 
+     * @param mixed $version versão do cache.
+     * @return string
+     */
+	private function GetVersion($version){
+		$version = !is_null($version) ? $version : $this->cfg['version'];
+		return !is_null($version) ? $this->Filter($version) : '';
 	}
 }